Move SummaryBoxHeading to its own file
